% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{atm_init}
\alias{atm_init}
\title{Initializes the Processed Additive Predictor for ATMs}
\usage{
atm_init(atmnr, h1nr)
}
\arguments{
\item{atmnr, h1nr}{positions of the atm and h1 formula}
}
\value{
returns a subnetwork_init function with pre-defined arguments
}
\description{
Initializes the Processed Additive Predictor for ATMs
}
