% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{ensemble.deeptrafo}
\alias{ensemble.deeptrafo}
\title{Deep ensembling for deeptrafo models}
\usage{
\method{ensemble}{deeptrafo}(
  x,
  n_ensemble = 5,
  reinitialize = TRUE,
  mylapply = lapply,
  verbose = FALSE,
  patience = 20,
  plot = TRUE,
  print_members = TRUE,
  stop_if_nan = TRUE,
  save_weights = TRUE,
  callbacks = list(),
  save_fun = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{"deeptrafo"}}

\item{n_ensemble}{numeric; number of ensemble members to fit}

\item{reinitialize}{logical; if \code{TRUE} (default), model weights are
initialized randomly prior to fitting each member. Fixed weights are
not affected}

\item{mylapply}{lapply function to be used; defaults to \code{lapply}}

\item{verbose}{whether to print training in each fold}

\item{patience}{number of patience for early stopping}

\item{plot}{whether to plot the resulting losses in each fold}

\item{print_members}{logical; print results for each member}

\item{stop_if_nan}{logical; whether to stop CV if NaN values occur}

\item{save_weights}{logical, whether to save weights of the ensemble.}

\item{callbacks}{a list of callbacks used for fitting}

\item{save_fun}{function applied to the model in each fold to be stored in
the final result}

\item{...}{further arguments passed to \code{object$fit_fun}}

\item{cv_folds}{an integer if list with train and test data sets}
}
\value{
ensemble of \code{"deeptrafo"} models with list of training histories
    and fitted weights included in \code{ensemble_results}
}
\description{
Deep ensembling for deeptrafo models
}
\examples{
dat <- data.frame(y = rnorm(100), x = rnorm(100))
m <- deeptrafo(y ~ 0 + x, data = dat)
ens <- ensemble(m, n_ensemble = 2)
coef(ens)

}
