% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{from_preds_to_trafo}
\alias{from_preds_to_trafo}
\title{Define Predictor of Transformation Model}
\usage{
from_preds_to_trafo(
  atm_toplayer = function(x) layer_dense(x, units = 1L),
  split = 1L,
  const_ia = NULL,
  ordered = FALSE
)
}
\arguments{
\item{atm_toplayer}{function to be applied on top of the transformed lags}

\item{split}{see \code{split_between_h1_h2} from \code{?deeptrafo}}

\item{const_ia}{see \code{addconst_interaction} from \code{?deeptrafo}
\code{deepregression}}
}
\value{
a function of list_pred_param returning a list of output tensors
that is passed to \code{model_fun} of \code{deepregression}
}
\description{
Define Predictor of Transformation Model
}
