% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_minval}
\alias{get_minval}
\title{Function to return the minval term}
\usage{
get_minval(x)
}
\arguments{
\item{x}{the fitted deeptrafo object}
}
\description{
Function to return the minval term
}
\details{
This value is only available if \code{addconst_interaction}
was specified in the model call.
}
