% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{h1_init}
\alias{h1_init}
\title{Initializes the Processed Additive Predictor for TM's Interaction}
\usage{
h1_init(yterms, h1pred)
}
\arguments{
\item{yterms, h1pred}{positions of the left and right RWT term}
}
\value{
returns a subnetwork_init function with pre-defined arguments
}
\description{
Initializes the Processed Additive Predictor for TM's Interaction
}
