% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{h1prime_init}
\alias{h1prime_init}
\title{Initializes the Processed Additive Predictor for TM's Interaction}
\usage{
h1prime_init(h1primenr, h1nr)
}
\arguments{
\item{h1primenr, h1nr}{positions of the h1 and h1prime formula}
}
\value{
returns a subnetwork_init function with pre-defined arguments
}
\description{
Initializes the Processed Additive Predictor for TM's Interaction
}
