% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{interaction_init}
\alias{interaction_init}
\title{Initializes the Processed Additive Predictor for TM's Interaction}
\usage{
interaction_init(
  pp,
  deep_top = NULL,
  orthog_fun = orthog_tf,
  split_fun = split_model,
  param_nr = 2
)
}
\arguments{
\item{pp}{processed predictor list from \code{processor}}

\item{deep_top}{keras layer if the top part of the deep network after orthogonalization
is different to the one extracted from the provided network}

\item{orthog_fun}{function used for orthogonalization}

\item{split_fun}{function to split the network to extract head}

\item{param_nr}{integer number for the distribution parameter}
}
\value{
returns a list of input and output for this additive predictor
}
\description{
Initializes the Processed Additive Predictor for TM's Interaction
}
