% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.deeptrafo}
\alias{logLik.deeptrafo}
\title{Log-likelihood method for deeptrafo objects}
\usage{
\method{logLik}{deeptrafo}(
  object,
  y = NULL,
  newdata = NULL,
  convert_fun = function(x, ...) -sum(x, ...),
  ...
)
}
\arguments{
\item{object}{deeptrafo object;}

\item{y}{vector; optional response}

\item{newdata}{data.frame; optional new data}

\item{convert_fun}{function; applied to the log-likelihood values of all observations}

\item{...}{currently not used}
}
\description{
Log-likelihood method for deeptrafo objects
}
