% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nll_count}
\alias{nll_count}
\title{Negative log-likelihood for count transformation models}
\usage{
nll_count(base_distribution = "logistic")
}
\arguments{
\item{base_distribution}{Base/error/target distribution. A parameter-free
distribution which defines the interpretational scale of the transformation
function in neural network transformation models. Can be character or
a \code{tfd_distribution}.}
}
\value{
A function for computing the negative log-likelihood of a
    neural network transformation model with count response.
}
\description{
Negative log-likelihood for count transformation models
}
