% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nll_ordinal}
\alias{nll_ordinal}
\title{Negative log-likelihood of an ordinal transformation model}
\usage{
nll_ordinal(base_distribution = "logistic")
}
\arguments{
\item{base_distribution}{Base/error/target distribution. A parameter-free
distribution which defines the interpretational scale of the transformation
function in neural network transformation models. Can be character or
a \code{tfd_distribution}.}
}
\value{
A function for computing the negative log-likelihood of a
    neural network transformation model with ordinal response.
}
\description{
Negative log-likelihood of an ordinal transformation model
}
