% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nll_surv}
\alias{nll_surv}
\title{Negative log-likelihood for potentially right-censored survival responses}
\usage{
nll_surv(base_distribution)
}
\arguments{
\item{base_distribution}{Base/error/target distribution. A parameter-free
distribution which defines the interpretational scale of the transformation
function in neural network transformation models. Can be character or
a \code{tfd_distribution}.}
}
\value{
A function for computing the negative log-likelihood of a
    neural network transformation model with survival response.
}
\description{
Negative log-likelihood for potentially right-censored survival responses
}
