% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias-deeptrafo.R
\name{ontram}
\alias{ontram}
\title{Ordinal neural network transformation models}
\usage{
ontram(
  response,
  intercept = NULL,
  shift = NULL,
  shared = NULL,
  data,
  lag_formula = NULL,
  response_type = "ordered",
  order = get_order(response_type, data[[all.vars(response)[1]]]),
  addconst_interaction = 0,
  family = "logistic",
  monitor_metrics = NULL,
  trafo_options = trafo_control(order_bsp = order, response_type = response_type),
  ...
)
}
\arguments{
\item{response}{formula for the response; e.g., \code{~ y}}

\item{intercept}{formula for the intercept function; e.g., \code{~ x},
for which interacting bases with the response will be set up}

\item{shift}{formula for the shift part of the model; e.g., \code{~ s(x)}}

\item{shared}{formula for sharing weights between predictors in the intercept
and shift part of the model}

\item{data}{\code{list} or \code{data.frame} containing both structured and
unstructured data}

\item{lag_formula}{formula representing the optional predictor for lags of the
response as defined for auto-regressive transformation models (ATMs)}

\item{response_type}{character; type of response. One of \code{"continuous"},
\code{"survival"}, \code{"count"}, \code{"ordered"}}

\item{order}{integer; order of Bernstein polynomials or number of levels for
ordinal responses}

\item{addconst_interaction}{positive constant;
a constant added to the additive predictor of the interaction term.
If \code{NULL}, terms are left unchanged. If 0 and predictors have
negative values in their design matrix, the minimum value of all predictors
is added to ensure positivity. If > 0, the minimum value plus the
\code{addconst_interaction} is added to each predictor in the interaction
term.}

\item{family}{\code{tfd_distribution} or character; the base distribution for
transformation models. If character, can be \code{"normal"}, \code{"logistic"}
or \code{"gumbel"}.}

\item{trafo_options}{options for transformation models such as the basis
function used}

\item{...}{Arguments passed to \code{deepregression}}
}
\value{
See return statement of \code{\link[deeptrafo]{deeptrafo}}
}
\description{
Ordinal neural network transformation models
}
\examples{
df <- data.frame(y = ordered(sample.int(6, 50, TRUE)), x = rnorm(50))
m <- ontram(response = ~ y, shift = ~ x, data = df)
fit(m, epochs = 2L)

}
\references{
Kook, L. & Herzog, L., Hothorn, T., Dürr, O., & Sick, B. (2022).
    Deep and interpretable regression models for ordinal outcomes.
    Pattern Recognition, 122, 108263. DOI 10.1016/j.patcog.2021.108263
}
