% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.deeptrafo}
\alias{predict.deeptrafo}
\title{Predict based on a deeptrafo object}
\usage{
\method{predict}{deeptrafo}(object, newdata = NULL, which = NULL, cast_float = FALSE, ...)
}
\arguments{
\item{object}{a deeptrafo model}

\item{newdata}{optional new data, either data.frame or list}

\item{...}{not used atm}
}
\value{
returns a function with two parameters: the actual response
and \code{type} in \code{c('trafo', 'pdf', 'cdf', 'interaction')}
determining the returned value
}
\description{
Predict based on a deeptrafo object
}
