% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.deeptrafo}
\alias{print.deeptrafo}
\title{Print method for deeptrafo objects}
\usage{
\method{print}{deeptrafo}(x, print_model = FALSE, print_coefs = TRUE, with_baseline = FALSE, ...)
}
\arguments{
\item{x}{deeptrafo object}

\item{print_model}{logical; print keras model}

\item{print_coefs}{logical; print coefficients}

\item{...}{currently not used}
}
\description{
Print method for deeptrafo objects
}
