% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.deeptrafo}
\alias{simulate.deeptrafo}
\title{Simulate method for deeptrafo objects}
\usage{
\method{simulate}{deeptrafo}(object, newdata = NULL, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{\code{"deeptrafo"} object}

\item{newdata}{list or data.frame; optional new data}

\item{nsim}{number of simulations; defaults to 1}

\item{seed}{seed for generating samples; defaults to \code{NULL}}

\item{...}{further arguments to \link[predict.deeptrafo]{predict.deeptrafo}}
}
\description{
Simulate method for deeptrafo objects
}
