% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.deeptrafo}
\alias{summary.deeptrafo}
\title{Summary method for deeptrafo objects}
\usage{
\method{print}{deeptrafo}(object, ...)
}
\arguments{
\item{object}{deeptrafo object}

\item{...}{further arguments supplied to \code{print.deeptrafo}}
}
\description{
Summary method for deeptrafo objects
}
