% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deeptrafo}
\alias{deeptrafo}
\title{Deep Conditional Transformation Models}
\usage{
deeptrafo(
  formula,
  data,
  lag_formula = NULL,
  response_type = get_response_type(data[[all.vars(fml)[1]]]),
  order = get_order(response_type, data[[all.vars(fml)[1]]]),
  addconst_interaction = 0,
  family = "logistic",
  monitor_metrics = NULL,
  trafo_options = trafo_control(order_bsp = order, response_type = response_type),
  return_data = FALSE,
  ...
)
}
\arguments{
\item{formula}{Formula specifying the response, interaction, shift terms
as \code{response | interacting ~ shifting}.}

\item{data}{Named \code{list} or \code{data.frame} which may contain both
structured and unstructured data.}

\item{lag_formula}{Optional formula for lags of the response in
auto-regressive transformation models (ATMs).}

\item{response_type}{Character; type of response. One of \code{"continuous"},
\code{"survival"}, \code{"count"}, or \code{"ordered"}. If not supplied
manually it is determined by the first entry in \code{data[[response]]}.}

\item{order}{Integer; order of the response basis. Default 10 for Bernstein
basis or number of levels minus one for ordinal responses.}

\item{addconst_interaction}{Positive constant;
a constant added to the additive predictor of the interaction term.
If \code{NULL}, terms are left unchanged. If 0 and predictors have
negative values in their design matrix, the minimum value of all predictors
is added to ensure positivity. If > 0, the minimum value plus the
\code{addconst_interaction} is added to each predictor in the interaction
term. This ensures a monotone non-decreasing transformation function in
the response when using (tensor product) spline bases in the interacting
term.}

\item{family}{A \code{tfd_distribution} or character; the base distribution for
transformation models. If character, can be \code{"normal"}, \code{"logistic"},
\code{"gumbel"} or \code{"gompertz"}.}

\item{monitor_metrics}{See \code{\link[deepregression]{deepregression}}}

\item{trafo_options}{Options for transformation models such as the basis
function used, see \code{\link[deeptrafo]{trafo_control}} for more details.}

\item{return_data}{Include full data in the returned object. Defaults to
\code{FALSE}. Set to \code{TRUE} if inteded to use
\code{\link[stats]{simulate}} afterwards.}

\item{...}{Additional arguments passed to \code{deepregression}}
}
\value{
An object of class \code{c("deeptrafo", "deepregression")}
}
\description{
Deep Conditional Transformation Models
}
\details{
\code{deeptrafo} is the main function for setting up neural network
    transformation models and is called by all aliases for the more special
    cases (see e.g. \code{\link[deeptrafo]{ColrNN}}). The naming convention
    of the aliases follow the 'tram' package (see e.g. \code{\link[tram]{Colr}})
    and add the suffix "NN" to the function name.
}
\examples{
data("wine", package = "ordinal")
wine$noise <- rnorm(nrow(wine))
fml <- rating ~ 0 + temp
m <- deeptrafo(fml, wine, family = "logistic", monitor_metric = NULL, return_data = TRUE)
m \%>\% fit(epochs = 20, batch_size = nrow(wine))
coef(m, which_param = "interacting")
coef(m, which_param = "shifting")

}
