% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{from_preds_to_trafo}
\alias{from_preds_to_trafo}
\title{Define Predictor of Transformation Model}
\usage{
from_preds_to_trafo(
  atm_toplayer = function(x) layer_dense(x, units = 1L, name = "atm_toplayer"),
  const_ia = NULL,
  ...
)
}
\arguments{
\item{atm_toplayer}{Function to be applied on top of the transformed lags.}

\item{const_ia}{See \code{addconst_interaction} in \code{\link[deeptrafo]{deeptrafo}}
or \code{\link[deepregression]{deepregression}}.}

\item{...}{For compatibility with 'deepregression'}
}
\value{
A function of \code{list_pred_param} returning a list of output tensors
that is passed to \code{model_fun} of \code{deepregression}
}
\description{
Define Predictor of Transformation Model
}
\details{
Not intended to be used directly by the end user.
}
