% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{h1_init}
\alias{h1_init}
\title{Initializes the Processed Additive Predictor for TM's Interaction}
\usage{
h1_init(yterms, h1pred, add_const_positiv = 0)
}
\arguments{
\item{yterms}{Terms for the response}

\item{h1pred}{Interacting predictor}

\item{add_const_positiv}{Shift basis for the predictors to be strictly
positive}
}
\value{
returns a subnetwork_init function with pre-defined arguments
}
\description{
Initializes the Processed Additive Predictor for TM's Interaction
}
