% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.deeptrafo}
\alias{plot.deeptrafo}
\alias{coef.deeptrafo}
\alias{predict.deeptrafo}
\alias{fitted.deeptrafo}
\title{Generic methods for deeptrafo objects}
\usage{
\method{plot}{deeptrafo}(
  x,
  which = NULL,
  which_param = "h1",
  only_data = FALSE,
  grid_length = 40,
  ...
)

\method{coef}{deeptrafo}(object, which_param = "h1", type = NULL, ...)

\method{predict}{deeptrafo}(object, newdata = NULL, ...)

\method{fitted}{deeptrafo}(object, newdata = NULL, batch_size = NULL, convert_fun = as.matrix, ...)
}
\arguments{
\item{x}{deeptrafo object}

\item{which}{which effect to plot, default selects all.}

\item{which_param}{integer, indicating for which distribution parameter
coefficients should be returned (default is first parameter)}

\item{only_data}{logical, if TRUE, only the data for plotting is returned}

\item{grid_length}{the length of an equidistant grid at which a two-dimensional function
is evaluated for plotting.}

\item{...}{not used atm}

\item{object}{a deeptrafo model}

\item{type}{either NULL (all types of coefficients are returned),
"linear" for linear coefficients or "smooth" for coefficients of;
Note that \code{type} is currently not used for \code{"h1"}}

\item{newdata}{optional new data, either data.frame or list}

\item{batch_size}{integer; optional, useful if data is too large}

\item{convert_fun}{function; to convert the TF tensor}

\item{eval_grid}{logical; should plot be evaluated on a grid}
}
\value{
returns a function with two parameters: the actual response
and \code{type} in \code{c('trafo', 'pdf', 'cdf', 'interaction')}
determining the returned value

returns a function with two parameters: the actual response
and \code{type} in \code{c('trafo', 'pdf', 'cdf', 'interaction')}
determining the returned value
}
\description{
Generic methods for deeptrafo objects
}
