% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.deeptrafo}
\alias{plot.deeptrafo}
\title{Generic functions for deeptrafo models}
\usage{
\method{plot}{deeptrafo}(
  x,
  which = NULL,
  which_param = 1,
  plot = TRUE,
  grid_length = 40,
  eval_grid = FALSE,
  ...
)
}
\arguments{
\item{x}{deepregression object}

\item{which}{which effect to plot, default selects all.}

\item{which_param}{integer, either 1 or 2.
1 corresponds to the shift term, 2 to the interaction term.}

\item{plot}{logical, if FALSE, only the data for plotting is returned}

\item{grid_length}{the length of an equidistant grid at which a two-dimensional function
is evaluated for plotting.}

\item{eval_grid}{logical; should plot be evaluated on a grid}

\item{...}{further arguments, passed to fit, plot or predict function}
}
\description{
Generic functions for deeptrafo models
}
