% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{neg_ll_trafo}
\alias{neg_ll_trafo}
\title{Generic negative log-likelihood of a transformation model}
\usage{
neg_ll_trafo(base_distribution)
}
\arguments{
\item{base_distribution}{base or error distribution}
}
\value{
a function for the negative log-likelihood with outcome \code{y}
and transformation model \code{model}. The transformation model is represented
by a list of two, with first element a list of model outputs
that are summed up and evaluated with the log-probability of the
\code{basis_dist}, and second element a single-column tensor
representing the determinant of the Jacobian and transformed
using the log
}
\description{
Generic negative log-likelihood of a transformation model
}
