% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nll}
\alias{nll}
\title{Generic negative log-likelihood for transformation models}
\usage{
nll(latent_distr)
}
\arguments{
\item{latent_distr}{Target distribution, character or
\code{tfd_distribution}. If character, can be either "logistic",
"normal", "gumbel", "gompertz".}
}
\value{
A function for computing the negative log-likelihood of a
    neural network transformation model with generic response.
}
\description{
Generic negative log-likelihood for transformation models
}
