% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{trafo_control}
\alias{trafo_control}
\title{Options for transformation models}
\usage{
trafo_control(
  order_bsp = 10L,
  support = function(y) range(y),
  y_basis_fun = NULL,
  y_basis_fun_prime = NULL,
  penalize_bsp = 0,
  order_bsp_penalty = 2,
  tf_bsps = FALSE,
  response_type = c("continuous", "ordered", "survival", "count"),
  atm_toplayer = function(x) layer_dense(x, units = 1L)
)
}
\arguments{
\item{order_bsp}{the order of Bernstein polynomials in case \code{y_basis_fun}
is a Bernstein polynomial defined by \code{eval_bsp}}

\item{support}{a function yielding a vector with two elements, namely
the support for the basis of y;}

\item{y_basis_fun, y_basis_fun_prime}{basis functions for y transformation}

\item{penalize_bsp}{scalar value > 0; amount of penalization of Bernstein polynomials}

\item{order_bsp_penalty}{integer; order of Bernstein polynomial penalty. 0 results in a
penalty based on integrated squared second order derivatives, values >= 1 in difference
penalties}

\item{tf_bsps}{logical; whether to use a TensorFlow implementation of the Bernstein polynomial
functions}

\item{response_type}{character; type of response can be continuous, ordered,
survival, or count}

\item{atm_toplayer}{function; a function specifying the layer on top of ATM lags}
}
\description{
Options for transformation models
}
